+++ src-docs/src/views/flyout/flyout_complicated.js
@@ -226,7 +226,6 @@ export default () => {
           <EuiPopover
             isOpen={isExpressionOpen}
             closePopover={() => setIsExpressionOpen(false)}
-            ownFocus={true}
             button={
               <EuiExpression
                 description="expression"