+++ src-docs/src/views/form_layouts/inline_popover.js
@@ -108,7 +108,6 @@ export default () => {
     <div>
       <EuiPopover
         id="inlineFormPopover"
-        ownFocus
         button={button}
         isOpen={isPopoverOpen}
         closePopover={closePopover}>
@@ -117,7 +116,6 @@ export default () => {
       &emsp;
       <EuiPopover
         id="formPopover"
-        ownFocus
         button={button2}
         isOpen={isPopover2Open}
         closePopover={closePopover2}>