+++ src/components/popover/popover.tsx
@@ -43,7 +43,7 @@ import {
 
 import { EuiScreenReaderOnly } from '../accessibility';
 
-import { EuiPanel, PanelPaddingSize } from '../panel';
+import { EuiPanel, PanelPaddingSize, EuiPanelProps } from '../panel';
 
 import { EuiPortal } from '../portal';
 
@@ -145,6 +145,10 @@ export interface EuiPopoverProps {
    * Standard DOM `style` attribute. Passed to the EuiPanel
    */
   panelStyle?: CSSProperties;
+  /**
+   * Object of props passed to EuiPanel
+   */
+  panelProps?: Omit<EuiPanelProps, 'style'>;
   panelRef?: RefCallback<HTMLElement | null>;
   popoverRef?: Ref<HTMLDivElement>;
   /**
@@ -309,7 +313,7 @@ type PropsWithDefaults = Props & {
 export class EuiPopover extends Component<Props, State> {
   static defaultProps: Partial<PropsWithDefaults> = {
     isOpen: false,
-    ownFocus: false,
+    ownFocus: true,
     anchorPosition: 'downCenter',
     panelPaddingSize: 'm',
     hasArrow: true,
@@ -458,7 +462,7 @@ export class EuiPopover extends Component<Props, State> {
 
     // for each child element of `this.panel`, find any transition duration we should wait for before stabilizing
     const { durationMatch, delayMatch } = Array.prototype.slice
-      .call(this.panel ? this.panel.children : [])
+      .call(this.panel ? [this.panel, ...Array.from(this.panel.children)] : [])
       .reduce(
         ({ durationMatch, delayMatch }, element) => {
           const transitionTimings = getTransitionTimings(element);
@@ -599,8 +603,6 @@ export class EuiPopover extends Component<Props, State> {
           ? anchorBoundingBox.left
           : left,
       zIndex,
-      // Adding `will-change` to reduce risk of a blurry animation in Chrome 86+
-      willChange: 'transform, opacity',
     };
 
     const willRenderArrow = !this.props.attachToAnchor && this.props.hasArrow;
@@ -663,6 +665,7 @@ export class EuiPopover extends Component<Props, State> {
       closePopover,
       panelClassName,
       panelPaddingSize,
+      panelProps,
       panelRef,
       panelStyle,
       popoverRef,
@@ -701,7 +704,8 @@ export class EuiPopover extends Component<Props, State> {
       { 'euiPopover__panel-isOpen': this.state.isOpening },
       { 'euiPopover__panel-noArrow': !hasArrow || attachToAnchor },
       { 'euiPopover__panel-isAttached': attachToAnchor },
-      panelClassName
+      panelClassName,
+      panelProps?.className
     );
 
     let panel;
@@ -756,6 +760,7 @@ export class EuiPopover extends Component<Props, State> {
               !ownFocus || !this.state.isOpenStable || this.state.isClosing
             }>
             <EuiPanel
+              {...(panelProps as EuiPanelProps)}
               panelRef={this.panelRef}
               className={panelClasses}
               hasShadow={false}
@@ -767,7 +772,13 @@ export class EuiPopover extends Component<Props, State> {
               aria-labelledby={ariaLabelledBy}
               aria-modal="true"
               aria-describedby={ariaDescribedby}
-              style={this.state.popoverStyles}>
+              style={{
+                ...this.state.popoverStyles,
+                // Adding `will-change` to reduce risk of a blurry animation in Chrome 86+
+                willChange: !this.state.isOpenStable
+                  ? 'transform, opacity'
+                  : undefined,
+              }}>
               <div className={arrowClassNames} style={this.state.arrowStyles}>
                 {arrowChildren}
               </div>