+++ CHANGELOG.md
@@ -1,14 +1,50 @@
 ## [`master`](https://github.com/elastic/eui/tree/master)
 
+- Added `indexRuntime` glyph in `EuiIcon` ([#4650](https://github.com/elastic/eui/pull/4650))
+- Added `iconType`, `iconColor`, and `iconSize` props to `EuiAvatar` ([#4620](https://github.com/elastic/eui/pull/4620))
+- Added `'plain'` and `null` as `color` options of `EuiAvatar` ([#4620](https://github.com/elastic/eui/pull/4620))
+
+## [`31.11.0`](https://github.com/elastic/eui/tree/v31.11.0)
+
+- Added `EuiNotificationEvent` component ([#4513](https://github.com/elastic/eui/pull/4513))
+- Added `euiAnimation()` method on the `EuiHeaderSectionItemButton` ref ([#4513](https://github.com/elastic/eui/pull/4513))
+- Made `description` prop of `EuiCard` optional ([#4546](https://github.com/elastic/eui/pull/4582))
+
+## [`31.10.0`](https://github.com/elastic/eui/tree/v31.10.0)
+
+- Added `panelProps` to `EuiPopover` ([#4573](https://github.com/elastic/eui/pull/4573))
+- Updated the default of the `EuiPopover`s `ownFocus` prop from `false` to `true` ([#4551](https://github.com/elastic/eui/pull/4551))
+- Added `src` prop to `EuiImage` as an alternative to `url` ([#4611](https://github.com/elastic/eui/pull/4611))
+- Added `EuiSplitPanel` component ([#4539](https://github.com/elastic/eui/pull/4539))
+
+**Bug fixes**
+
+- Fixed hover effect of nested clickable rows in `EuiBasicTable` ([#4566](https://github.com/elastic/eui/pull/4566))
+- Fixed visual bug in drag&drop sections when nested in an popover ([#4590](https://github.com/elastic/eui/pull/4590))
+- Fixed an errant export of `EuiSideNavProps` type from JS code ([#4604](https://github.com/elastic/eui/pull/4604))
+- Fixed misaligned `EuiComboBox` options list ([#4607](https://github.com/elastic/eui/pull/4607))
+- Fixed missing `forwardRef` on `EuiHeaderSectionItemButton` ([#4631](https://github.com/elastic/eui/pull/4631))
+
+## [`31.9.1`](https://github.com/elastic/eui/tree/v31.9.1)
+
+**Bug fixes**
+
+- Fixed an errant export of two non-existant values ([#4597](https://github.com/elastic/eui/pull/4597))
+
+## [`31.9.0`](https://github.com/elastic/eui/tree/v31.9.0)
+
+- Added `EuiComboBoxOptionOption` prop to `EuiComboBox` props table ([#4563](https://github.com/elastic/eui/pull/4563))
 - Allowed dynamically changing the `direction` prop on `EuiResizableContainer` ([#4557](https://github.com/elastic/eui/pull/4557))
 - Exported `useIsWithinBreakpoints` hook ([#4557](https://github.com/elastic/eui/pull/4557))
+- Added focus to `EuiForm` error `EuiCallout` ([#4497](https://github.com/elastic/eui/pull/4497))
+- Added a `display` and `size` props to `EuiButtonIcon` ([#4466](https://github.com/elastic/eui/pull/4466))
 
 **Bug fixes**
 
 - Fixed the return type of `getDefaultEuiMarkdownUiPlugins` ([#4567](https://github.com/elastic/eui/pull/4567))
 - Fixed inverse handling of boolean sorting in `EuiDataGrid` ([#4561](https://github.com/elastic/eui/pull/4561))
 
-## Feature: EuiPageTemplate ([#4517](https://github.com/elastic/eui/pull/4517))
+### Feature: EuiPageTemplate ([#4517](https://github.com/elastic/eui/pull/4517))
 
 - Added new `EuiPageTemplate` component as a shortcut for creating the different types of page layout patterns
 - Added props `grow` and `direction` to `EuiPage`